Fix handling for health of empty indices in ES7.
Before this change, the ES7 implementation of `ClusterAdapterES7#health`
was inconsistent with the one for ES6. When an empty list was passed as
an argument, an exception was thrown, instead of a valid response being
returned. The method being called with an empty list can happen if it is
called from `Cluster#health`, which retrieves all existing indices for
an index set before using `Cluster#allIndexWildcards`, which returns an
empty list if these have not been created yet. (E.g. as the consequence
of a race condition when starting Graylog from scratch/with an empty
ES). From what was seen during fixing this issue, the only relevant
consequence is that `IndexRangesMigrationPeriodical` throws an exception
and will be rerun during next startup. Usually, this is not an issue, as
there are no index ranges to be migrated at this point anyway.

This change is now making the ES7 implementation of
`ClusterAdapter#health` consistent with the one for ES6, by:

  - Returning `Optional.empty()` if the list of indices is not empty and
not all of them exist.
  - Passing index options that do not return an error if none of the
indices exist.

In addition, the read timeout is configured for the request just as in
`ClusterAdapterES6#health`.