Stop pivot config form submit event propagation. (#9085)
* Stop pivot config form submit event propagation.

Before this change, when configuring the settings for a row/column pivot
in the widget edit mode, submitting the form resulted in a double
submit. First, the pivot config form is submitted and its config is
changed. Then, the same submit event is propagated to the overlying
widget edit form (as defined in `SearchBarForm`). The `onSubmit` handler
in there is submitting the current widget again, but with stale pivot
config settings, resulting in a mismatching view/search constellation.
This results then in search results not being resolvable, as the widget
mapping is based on a stale search.

This change is now preventing the pivot config form submit event to be
propagated, therefore limiting its effect to the initial submit only.

* Adding test case.