Add ownership check to saved search/dashboard permission checks. (#9067)
* Respect ownership of saved search/dashboards.

Before this change, only user permissions were considered when
permission checks for actions operating on saved searches/dashboards
were performed. This lead to migrated saved searches having an owner,
but these owners were not able to access their migrated saved searches.

With this change, any action is also considering the owner of a saved
search/dashboard before checking permissions. An owner of a saved
search/dashboard is able to perform any action on it. If the current
user is not the owner, user permissions and view sharings are checked,
just as before.

* Adding owner check for dashboards.

* Adding tests.