Sync cloud branch with current 3.3. branch (#9104)
* Bump package.json version to 3.3.5

* [graylog2-server] prepare release 3.3.5

* Bump package.json version to 3.3.6-SNAPSHOT

* [graylog2-server] prepare for next development iteration

* Fix path traversal in swagger resource. (3.3) (#8988)

Prior to this change, the `/api-browser/.*` resource allowed to retrieve
static assets used by swagger. Some of these were prefixed with a
`global/` prefix, that needs to be stripped away before locating the
file. The string replacement used for this was used unbounded (replacing
every occurrence of the prefix, not just the first one), so it could be
used to create URLs containing `..`. The check if the route contains
`..` was performed _before_ the string replacement, so it could
effectively be used to evade it.

This change is:
  a) Replacing only the first occurrence of `global/`
  b) Performing the check if the route contains `..` _after_ the
replacement and _right before_ the file is tried to be opened.

Fixes #8986

(cherry picked from commit b42f85fbed25ee5446544697bfc06bcce218f7bf)

* Reset Consumes annotation to accept everything for (un)schedule (#8981) (#8993)

Since these endpoints don't accept a request body, we can just accept
any content-type. This unbreaks calling these endpoints with our swagger
API browser.

Co-authored-by: Bernd Ahlers <bernd@users.noreply.github.com>

* Adding title attr to Pipeline Title Link

* Remove permission check for Search Nav Item (3.3) (#8969)

* Remove permission check for Search Nav Item (#8917)

* Remove permission check for Search Nav Item

Prior to this change, the Search Menu Item was only displayd for admins
or users which have searches:absolute', 'searches:relative',
'searches:keyword' permissions.

The mentioned permission could only be add via API requests. Which left
the admin users as the only ones which could really see that Menu Item.

But a user could go to that page either via a link (on the streams page)
or by typing the url manually.

This change will now simply remove this check and always display the
menu item. The search still will only work if the user has read
permissions to a stream.

* Change default start page to search

the start page was set depending on the
search permissions a user has.

Now the start page will always default to Search Page

(cherry picked from commit ca224b8dc8b39a4c7005cf20f7a34a853b59be25)

* Adjusting item count.

* Fixing/suppressing linter hints.

* Suppressing linter hints.

Co-authored-by: Konrad Merz <konrad@graylog.com>

* check if config is available before rendering page (#9018)

* Merge new notification type to 3.3 (#9020)

When outputs fail, we can't simply shut them down.  Thus, we need to give the users a clear message about what has happened.  This change adds a new notification type for outputs that are failing so we can do just that.

* Show better error message when user has no stream permissions. (3.3) (#9011)

* Show better error message when user has no stream permissions.

Before this change, the user was seeing a "Missing stream permissions"
page, that did not show a helpful error message if the user is not
missing the permission for a specific stream, but for any stream at all.

This change is now checking if the user is having any stream permissions
at all and shows a better message otherwise.

Fixes #8955.

* Fixing typo.

* Minor cleanups.

* Refresh streams store upon mounting provider.

* Include streams context in tests.

Co-authored-by: Dennis Oelkers <dennis@graylog.com>

* Submit widget query form before saving edit widget modal. (Backport of #8985 for 3.3) (#9043)

* Submit widget query form before saving edit widget modal. (#8985)

* Submit widget query form before saving edit widget modal.

Before this change, a successful change to the timerange/stream
selection/query of a widget on a dashboard required submitting the
widget's search after the change and before clicking "Save". This was
unintuitive and unnecessarily complicated.

This change is now submitting the widget's query form when the user
clicks "Save" and the form is dirty, not requiring the user to have
submitted the form manually before.

Fixes #7922.

* Submit only if dirty.

* Removing unnecessary form element around events checkbox.

* Updating snapshots.

* Removing usages of `widget` prop for `WidgetQueryControls`.

* Reusing `DEFAULT_TIMERANGE`.

* Include `dirty` in `useCallback` dependencies.

* Remove SearchBarForm validation which is not related to this backport and not part of 3.3

* Fix and update WidgetQueryControls test

* Update snapshots

* Fix EditWidgetFrameTest

Co-authored-by: Dennis Oelkers <dennis@graylog.com>

* Bump package.json version to 3.3.6

* [graylog2-server] prepare release 3.3.6

* Bump package.json version to 3.3.7-SNAPSHOT

* [graylog2-server] prepare for next development iteration

* Update browserslist DB. (#9052)

This change is solely updating the browserslist DB, so we are building
against the most recent database of target browsers.

* Stop pivot config form submit event propagation. (#9085)

* Stop pivot config form submit event propagation.

Before this change, when configuring the settings for a row/column pivot
in the widget edit mode, submitting the form resulted in a double
submit. First, the pivot config form is submitted and its config is
changed. Then, the same submit event is propagated to the overlying
widget edit form (as defined in `SearchBarForm`). The `onSubmit` handler
in there is submitting the current widget again, but with stale pivot
config settings, resulting in a mismatching view/search constellation.
This results then in search results not being resolvable, as the widget
mapping is based on a stale search.

This change is now preventing the pivot config form submit event to be
propagated, therefore limiting its effect to the initial submit only.

* Adding test case.

* Update dependencies with security issues (#9098)

* Bump jquery to 3.5.1

* Bump elliptic to 6.5.3

* Bump lodash to 4.17.20

* Bump serialize-javascript to 4.0.0

Since that is a transitive dependency, we update some packages depending
on it, specially terser-webpack-plugin.

Co-authored-by: Jenkins <garybot2@torch.sh>
Co-authored-by: Dennis Oelkers <dennis@graylog.com>
Co-authored-by: Linus Pahl <46300478+linuspahl@users.noreply.github.com>
Co-authored-by: Bernd Ahlers <bernd@users.noreply.github.com>
Co-authored-by: Kyle Knight <kyle@graylog.com>
Co-authored-by: Konrad Merz <konrad@graylog.com>
Co-authored-by: Rob Curtis <6466251+waab76@users.noreply.github.com>