Handle grants update with different capabilies
Change DBGrantService.ensure() to allow updating grants
to a higher capability, but not to a lower one.

This is used by migrations that could try to create
different grants for the same target.
After all migrations are run, we will end up with the grant that
has the most powerful capability.