Store username in localStorage which gets returned by the backend on login
Currently the username entered in the login form is being stored in the localStorage. This login username can be different than the actual username, which gets returned by the backend on login.
We need to store the username provided by the backend, otherwise we can't load the user if these usernames vary.
This is not a problem with the default setup, but it can be problematic when an authentication service is configured.