Re-register authentication components.
In #9118 a change slipped in that removes the registration of
authentication components (it is moved in a future PR to the `AppRouter`
component). This leads to the "Change Provider Order" subsection of the
System/Authentication page to have blank table cells, as well as the
subnavigation on the left showing short identifiers.

This change is now putting the registration back in to fix this.