Fix route retrieval in `ConfirmLeaveDialog`. (#9145)
In #9118, a refactoring was performed that removes passing the current
route as a prop to `ConfirmLeaveDialog`. It was erroneously assumed that
`withRouter` also supplies the `route` object. This was effectively
breaking the dialog that is shown in production when the current route
is left (by using the navigation), the current browser tab is closed or
reloaded.

This change is now retrieving the current route from the `router.routes`
object (the current child route is the last object) and uses it to
register the hook called when the current route is left.

In the future (with react-router v5), another refactoring is performed
replacing the imperative logic from `ConfirmLeaveDialog` with the usage
of the `Prompt` component.