Fix JUnit5 vintage engine by updating to v5.7.0. (#9168)
In #9139 JUnit4 was updated to `v4.13.1` due to a security issue.
Unfortunately, this version number is breaking the version check in the
JUnit 5 vintage engine of v5.6.0 and causes all JUnit5 tests to be
silently ignored, although an exception during version parsing is
thrown.

This change is now updating JUnit5 to v5.7.0. Unfortunately,
testcontainers is using v5.6.2, effectively overriding the version for
`junit-jupiter-api`, breaking v5.7.0. Therefore, this change is also
updating testcontainers to v1.15.0-rc2 (unfortunately there is currently
no stable version using v5.7.0).