Fix plugin routes with `null` parent components.
Our application router allows plugins to register routes with a `null`
parent component, which allows them to be rendered outside of our
application chrome. This got broke in #9135, as react-router v5 is now
rendering all matching routes (unless `Switch` is used) and the scoping
of the `Switch` components and the plugin routes got messed up.

This change is now restoring the original behavior by

  a) placing the `RouterErrorBoundary` outside of the `Switch` clause.
  b) putting the application chrome components (`App`/`AppWith...`)
inside the catch-all route for non-null-parent components.