Rename EventNotification GRNType in frontend
Prior to this change, the dependency resolution which is using the
content pack mechanism tried to find the event notification type with
'event_notification' which is not a registered type in content pack.

This change will rename the type to `notification` so the dependency
resolution via content packs works again.