Downgrading webpack-cli to v3.3.12.
Recently, `webpack-cli` was upgraded to v4.x. It was found out, that it
contains a regression that causes failing builds (e.g. due to a missing
npm dependency) do not result in a non-zero error code, effectively
preventing build pipelines to abort in an error case. While this seems
to be fixed in a PR (webpack/webpack-cli#1971), it is not included in
any released version at this time.

Therefore this change is downgrading `webpack-cli` to the most recent
3.x version. As we do not rely on any changes in 4.x, we can safely go
back to have a version which is working as expected for the upcoming
release of Graylog 4.x. During the next development iteration we can
retry the upgrade of `webpack-cli`, hoping that these kinds of issues
had been ironed out by then.