Fix field types of current query if no stream is selected.
Before this change, the `FieldTypesStore` returns empty field types for
the current query, if the current query has no streams selected (which
results in all streams being queried).

This change is now fixing the `QueryFiltersStore` to return queries with
an empty filter as well, and fixes the `FieldTypesStore` to return all
field types for queries which have an empty filter.