Saved Search: Hide Save button for missing edit perissions
## Motivation
Prior to this change, the Save button was shown to a user which only had
view permissions to a Saved Search. If he clicked on that button the
result was an error page.

## Description
This change will hide the 'Save' and 'Save as' button and shows a
'Create New' button to make clear he has to create a new Search which he
can edit.