Add target as href to `LinkContainer` children.
Before this change, the `LinkContainer` component did add an `onClick`
handler to its children to perform a route change without reloading the
app. This lead to the element being created not having an `href` prop,
so clicking on it and opening it in a new tab did not work.

This change is adding an `href` prop containing the value of the `to`
prop to its children, so the html element contains a valid target.