Avoid using throw by displaying an Alert
## Motivation
Prior to this change, when we had an runtime error we rendered the error
page to the user even though he might have been able to workarround an
issue.

## Description
This change implements a ErrorAlert component which is closable and
which will display an runtime error or warning. It replaces the
warnings already implemented and does also replace the `throw` errors
for users and roles components.