Do not stop click events on `LinkContainer` if modifier-keys are used.
Before this change, all click events on a `LinkContainer` were stopped,
including click events with modifier keys (e.g. Ctrl) being held. This
prevents browser functionality like Ctrl+Click to open a link in a new
tab.

This change is adding a check to the generated `onClick` handler,
stopping all processing and stopping the event if a modifier was held
during clicking.