Add migration for legacy LDAP/AD configuration
This creates a migration to create a new authentication service based on
an existing legacy LDAP/AD configuration.

Since the legacy LDAP/AD configuration doesn't contain all information
we need for a new authentication service, we guess some config fields
and create a system notification to make sure an admin will notice it
and reviews and activates the migrated configuration.

Closes #9078