Disabling rebase action for github
This PR is disabling the '/rebase' action, as it might be vulnerable to command injection. It might be reenabled when it turns out to be safe or the vulnerability is fixed.

Details are available [here](https://bugs.chromium.org/p/project-zero/issues/detail?id=2070&can=2&q=&colspec=ID%20Type%20Status%20Priority%20Milestone%20Owner%20Summary&cells=ids).