Retry indexing message if primary shard is not active for ES6. (#9338)
Prior to this change, when a message is tried to be ingested into Elasticsearch when the primary shard of an index is unavailable, indexing will fail and the message will be turned into an indexing failure. Most probably, this is not what a user wants.
In most cases an unavailable primary shard is a) affecting most, if not all indices, b) is a temporary condition. These properties and the similarities to scenarios where a watermark threshold switches indices to read-only do justify a handling of this that attempts retries indefinitely until the condition is resolved.

While this prevents message loss, it can also cause journal overflow (and causing message loss at the other end of ingestion), while it buys operations time to fix the issue. Due to the watermark scenario, proper sizing of the journal is required anyway to prevent catastrophic scenarios where message indexing stops completely.

This change is extending parsing of ES errors returned during ingestion and categorizes the "primary shard not available" error in the same way as the "index r/o" error. Therefore, errors like this will be retried indefinitely, until the condition is resolved.