We put new selections to the top of selceted grantees
## Motivation
Prior to this change, if a share dialog has a lot of selected grantees
the pagination is shown. With the current ordering new entries to the
selection could be displayed somewhere in the pagination and a user
would not see that he had selected a new entry.

## Description
This change will put the new grantees to the beginning of the entity
share modal.