Make sure to use the OSS ES container for tests (#9394)
* Make sure to use the OSS ES container for tests

Due to the version constraint of "^6.0.0" we only used the OSS container
for ES 6.x.

Adjust ClientES7#existingTemplates to not use a wildcard ("*") argument
when creating the GetIndexTemplatesRequest object.
Using the wildcard argument with no existing index templates returns a
404 from ES and throws an exception.

This fixes the IndicesGetAllMessageFieldsES7IT test when switching to
the OSS container. The regular ES container is automatically creating
indices for x-pack features so there are always existing indices and
index templates. The OSS container doesn't create any indices and
templates by default and so the #existingTemplates call failed with a
404.

* Disable ES container reuse on Jenkins

This prevents a warning on every container startup.

* Remove default ES image selection

We don't support ES <6 anymore so we can always use the OSS image.