Create session when user can be authenticated based on trusted header (#9404)
* Request session creation when authentication for trusted header suceeded

When a subject is authenticated via a trusted header, we have to signal
to the session resource to create a session. This way, the initial
session validation request from the frontend will lead to the creation
of a session which can then be used for further requests.

* Return username instead of user ID in session

The main principal of the subject has been changed to the user id. The
response to the session validation request should however contain the
username so we have to load the user explicitly and get the name.