Prevent disabling your own user account
## Motivation
Prior to this change, a logged in user was able to disable himself (if
he had the necessary permissions).

## Description
This change will hide the menu item `Enable`/`Disable` for the user
himself.

## Also
Add a confirmation dialog for disabling user which explains that
disabling a user also stops its current session.