Add reset startpage button to profile section of user edit
## Motivation
Prior to this change, there was no way to reset a startpage for a user.

## Description
This change will add the most minimal solution by adding a Reset
Startpage button to the header of the profile section on the user edit page.

## Also
Fix setting the Set as Startpage menuitem for streams.