Fix scoping of current page in system messages list. (#9456)
During a refactoring, the scoping of the `currentPage` variable used to
specify the page of system messages which is currently fetched was
changed, effectively fixing it to its initial value. This leads to page
changes not being reflected for the periodical refresh.

This change is now adjusting the scope and using the value of the
current page which is up to date.