Fix index field types update (#9463) (#9468)
In #9177 we introduced a check to skip the periodical run when the
server is not in RUNNING state. This was added to fix exceptions during
server shutdown by preventing the periodical to run in that state.

Unfortunately this is will delay index field types update by an hour
when the periodical starts for the first time on server startup. The
periodical only runs once an hour by default.

This change is removing the RUNNING state check in the periodical's
run() method. We keep the check in the individual poller jobs. These are
running way more often.

This change also removes the shutdown hook that closes the ES client
during server shutdown. Closing the client connection can cause
exceptions during the server shutdown when the index field type poller
job (and others) run during shutdown and try to use a closed client.

Since all connections will be closed during process shutdown anyway, we
skip the closing for ES 7 to prevent exceptions. (we didn't do that for
the ES 6 client anyway)

(cherry picked from commit c46db0d8cc82f8fea2aac8197e47c90defb51eee)