Improve `IndicesIT` test by using delta when comparing dates. (#9457)
This change is improving one of the `IndicesIT` tests, as it has turned
out to be flaky when it is executed at the brink of a second. In this
case, the reference date (e.g. `2020-11-12T09:01:37.999Z`) is not equal
to the actual one (e.g. `2020-11-12T09:01:38.000Z`), even when ignoring
milliseconds. Therefore, this change is changing the assertion to check
if those two dates are close enough to each other with a delta of one
second.