Remove explicit shutdown for messageQueueReader
Originally, this shutdown didn't have any effect anyways, because the
object it was called upon was not the running service. That's because of
 the missing @Singleton annotation in the former JournalReader class.

During shutdown, the message queue reader will stop processing because
it listens to lifecycle events. At least the Kafka version does.

There is no need to change the behaviour now.