Lifecycle handling for Pulsar message queue reader (#9735)
* Add missing @Singleton annotation

To make sure that if the object is injected, it is actually the same
object as bound via the service binder.

* Remove explicit shutdown for messageQueueReader

Originally, this shutdown didn't have any effect anyways, because the
object it was called upon was not the running service. That's because of
 the missing @Singleton annotation in the former JournalReader class.

During shutdown, the message queue reader will stop processing because
it listens to lifecycle events. At least the Kafka version does.

There is no need to change the behaviour now.

* Fix comment

Interrupting the thread won't help, so just document the issue for now

* Extract lifecycle handling and share for Kafka and Pulsar readers

* Simplify lifecycle switch

* Add license header