Always short field sort icons in message table.
Before this change, the field sort icons used to change sorting of a
message table were shown in edit mode only. Most users are used to
change sorting outside of edit mode by clicking an icon next to the
field name in the table header.

This change is showing sorting icons outside of the edit mode too, but
not in non-interactive contexts.