Unblock JobSchedulerService on shutdown
The JobSchedulerService blocks in its run method until the server
reaches the running state. It cannot be unblocked even when interrupting
the execution thread. So if one of the other services fails during
startup, the service manager is unable to shutdown the
JobSchedulerService.

This change adds an interruptible implementation to wait for the server
reaching the RUNNING state. The new method is then used in the
JobSchedulerService. Upon shutdown, the execution thread of the
JobSchedulerService is interrupted so that the service can
properly terminate.