Retry read timeouts for Elasticsearch.
Before this change, the retry strategy used for ES requests did not
retry for instances of `SocketTimeoutException`, raised in case of read
timeouts. This resulted in exceptions being raised when individual nodes
were not responding and node discovery was not used. This was e.g. being
noticable when performing searches and 1 of 3 nodes was failing, so
every 3rd search request was timing out and returning an error instead
of being retried and successful and just taking longer.

This change is adding this class to the exceptions being retried, so any
request that goes out to a node which is not responding on time is
retried.