Fix csv export for shared dashboards.
## Motivation
Prior to this change, if a user had read access to a dashboard via role
or direct permission the csv export failed for this dashboard, since the
permission test was testing on 'VIEW' where the role and permission had 'DASHBOARD'.

## Description
This change will use a new funtion introduced just for this case.