Avoid using a stage number that is already used
Check if selected stage number already exists in pipeline, displaying an
error in that case and not allowing to submit the form. This will
prevent users of accidentally overriding one of their existing stages.

Additionally improve the initial suggestion for new stages. Before we
always displayed `0` when creating a stage, now we base that value in
the existing stages for the pipeline.

Fixes #6806