Supporting `from` and `to` properties in relative time ranges. (#9899)
* Supporting `from` and `to` properties in relative time ranges.

This change is implementing functionality to support an alternative way
to use relative time ranges. Instead of a `range` property that defines
the length of the range in seconds going backwards from now, it allows
defining `from` and `to` properties defining the endpoints of a range in
relation to `now`. This allows e.g. defining a range that goes from 5
minutes ago to 1 minute ago by specifying a relative range as:

```
{
  "type": "relative",
  "from": 300,
  "to": 60
}
```

Fixes #9898.

* Sanity checking presence and relation of `from`/`to`.

* Fixing missing brace.

* Handle all message ranges as before.