Sync cloud 4.0 (#9986)
* Fix order of tabs in dashboard after install (#9812)

## Motivation
Prior to this change, a dashboard installed by a content pack had a
tabs in a random order. Since we cannot change the order of the tabs
this is quite inconvinient for the user.

## Description
This change will do mutiple steps to fix that:
- Use a LinkedHashMap to ensure the order stays fix during content pack
  creation.
- Instead of using the serach QueryIds we use the views QueryIds which
  we previously fixed in the order.
- When loading a view and selecting the activeQuery we also use the
  view.state queryids instead of the search ones.

Fixes #8264

* Show timezone setting to the user (#9814)

* Display the timezone settings to the user

## Motivation
Prior to this change, the user was not able to change his own timezone
settings since they were hidden to him.

## Description
This change will show the settings to the user so he can edit them.

* Fix settings section test

* Bump urijs to 1.19.5 (#9874)

This provides a fix for the following security issue:
https://github.com/advisories/GHSA-3329-pjwv-fjpg

Backport of #9857 for 4.0

* Fix pattern used to identify parameters in search/dashboard query strings. (#9877) (#9887)

* Consolidating backend-agnostic query parsing/metadata creation.

* Fixing test.

* Adding test cases for bug.

* Adjusting regex for query string parser.

* Removing unused parameter.

Co-authored-by: Dennis Oelkers <dennis@graylog.com>

* Fix deserialization of incomplete AlertReceivers objects (#9785) (#9910)

The new paginated endpoint in StreamResource is using mongojack to
deserialize the database objects. In some setups the AlertReceivers
object in the database might not be complete.

Change AlertReceivers to use defaults for missing values.

Fixes #9637

(cherry picked from commit 1e456af53b5853182d769dbeb32ac0c9cf8a240c)

* Fix page size change selector in Content Pack list (#9875) (#9924)

The `PageSizeSelect` component passes an event as argument to its
`onChange` function but but we were expecting a string containing
a number. This change fixes that by extracting the value from the
received event.

Fixes #9629

Co-authored-by: Edmundo Alvarez <edmundo@graylog.com>

* Do not allow selecting same pipeline stage twice (#9916)

* Avoid using a stage number that is already used

 Check if selected stage number already exists in pipeline, displaying
 an error in that case and not allowing to submit the form. This will
 prevent users of accidentally overriding one of their existing stages.

 Additionally improve the initial suggestion for new stages. Before we
 always displayed `0` when creating a stage, now we base that value in
 the existing stages for the pipeline.

 Fixes #6806

* Do not submit form when using an existing stage

* Retry read timeouts for Elasticsearch (#9922)

* Retry read timeouts for Elasticsearch.

Before this change, the retry strategy used for ES requests did not
retry for instances of `SocketTimeoutException`, raised in case of read
timeouts. This resulted in exceptions being raised when individual nodes
were not responding and node discovery was not used. This was e.g. being
noticable when performing searches and 1 of 3 nodes was failing, so
every 3rd search request was timing out and returning an error instead
of being retried and successful and just taking longer.

This change is adding this class to the exceptions being retried, so any
request that goes out to a node which is not responding on time is
retried.

* Removing duplicate license header.

* Fix date format specifiers in index mapping templates (#9948)

* Fix date format specifiers in index mapping templates for ES7.

Before this change, the pre-ES7 date format specifiers were reused for
ES7 to define the format of date fields in index mapping templates.

Unfortunately, this results in warnings being returned in responses to
searches operating on date fields, due to a [breaking change introduced
in ES7](https://www.elastic.co/guide/en/elasticsearch/reference/current/breaking-changes-7.0.html#_joda_based_date_formatters_are_replaced_with_java_ones).

This change is now splitting up the date format specifiers and uses joda
syntax for ES6 and `java.time` syntax for ES7. In addition, the overly
lax exposition of the date format specifiers from the `Tools` class was
locked down to prevent them leaking into new places.

Fixes #9690.

* Adjusting tests/expected results.

* Adjusting single date format in GIM Mapping for v7 too.

* Prefixing date format for ES6 to disable deprecation notice.

As we are using a deprecated formatter according to the [ES breaking changes](https://www.elastic.co/guide/en/elasticsearch/reference/current/breaking-changes-7.0.html#_joda_based_date_formatters_are_replaced_with_java_ones),
this commit is adding the `8` prefix to prevent a warning.

* Prefixing legacy date formats in GIM mapping.

* Adjusting tests to use `8yyyy`.

* Fixing stream rule form state update on input change. (#9968) (#9975)

Before this change we were mutating this.state on input changes directly which resulted in a strange behaviour.
One result was that the stream rule form did not reset after creating a stream rule, even though the component is unmounting correctly.
Have a look at #9956 for more information.

* Fix SourceCode editor combined with pipeline rules (4.0) (#9976)

Co-authored-by: Gaya <gaya@theclevernode.com>

* Retrying bulk indexing if target alias has invalid target(s) (#9978)

* Adding test to check if indexing retries for invalid alias targets.

* Adding indexing listener to bulk indexing to improve tests.

* Suppressing specific warnings.

* Introducing `InvalidWriteTargetException`.

* Retrying bulk indexing upon `InvalidWriteTargetException`.

* Relaxing ES exception regex.

* Throw specific exception if parsed exception indices invalid write target.

* Fixing error parsing for ESS6, extracting target for exception.

* Reverting changes to `ElasticsearchException`.

* Adding related test case for ES exception parser.

* Fix implicit mutation of retryer builder when listener is present.

Fixes #9955
Fixes #9962

* Bump package.json version to 4.0.2

* [graylog2-server] prepare release 4.0.2

* Bump package.json version to 4.0.3-SNAPSHOT

* [graylog2-server] prepare for next development iteration

Co-authored-by: Konrad Merz <konrad@graylog.com>
Co-authored-by: Dennis Oelkers <dennis@graylog.com>
Co-authored-by: Bernd Ahlers <bernd@users.noreply.github.com>
Co-authored-by: Linus Pahl <linus@graylog.com>
Co-authored-by: Gaya <gaya@theclevernode.com>
Co-authored-by: Jenkins <garybot2@torch.sh>