Use classgraph library to improve JerseyService startup speed
This reduces the startup time of the JerseyService by 83% on my
development machine when running the production artifacts.
(3.5s vs 20.5s)

Since scanning package paths for resources is pretty slow in Jersey,
this replaces the usage of ResourceConfig#packages with
ResourceConfig#registerClasses.

We now scan for resource classes by using the classgraph library.