Provide default position in case the widget isn't found in positions map (#9841)
* Provide default position in case the widget isn't found in positions map

## Motivation
Prior to this change, a widget copied from a search was unable to be
copied to a different page since the position was missing in
the positions map.

## Description
This change will add a safty net in case the position is missing we
provide the widgets default size.

* Use AddNewWidgetsToPosition for CopyWidgetToDashboard

That way we can saftly assume that future copied widgets will not have a missing position

The NaN values in the test can be ignored, since the widgets are not
existing anymore.

* Fix linter warning

* Fix license header

* Clear positions from stale entries

* Provide defaultHeight/Width to test

* Update snapshots

* Refactor position calculation for new widgets

* Fix linter warnings

* Create new id when copying a widget

* Reuse old sizes if possible

* Rename last leftover import

* Fix linter

* Don't mix responsibilities

* Update snapshots