Adjustments for enterprise widget
## Motivation
Prior to this change, we handled the z-index for the sidebar in way that
it was hard to add new z-index for widgets.
Also we could not disable a icon button.

## Description
This change raise the z-index for the sidebar so widget have a little
space for their own z-index without jumping out of the overlay.
Add IconButton disable prop.