Don't execute a search when focusing or unfocusing a widget (#10053)
## Motivation
Prior to this change,  if the user focused or unfocused a widget a
search was executed, since the code assumed the streams filters changed.

## Description
This change will cleanup the filters we get from the backend (it
included title and other filters which where set to null).
Also to compare the new streams filter and the old one the need to be
from the same object type. So we enusre that the new streams filter is
also containing immutable list and not an array.