Improve build stability of `graylog-web-plugin` module.
In recent weeks, we have seen build instabilities for the
`graylog-web-plugin` module, where the build contains an empty `lib`
directory, no files from the `src` dir were transpiled, but `yarn
install` succeeded.

The causes for this were:

- The `preinstall` task erroneously catching build errors (the `||`
  clause was meant to suppress build errors when no `src` dir is
present, which should not happen anymore)

- Installing inconsistent `babel` dependencies. `@babel/preset-env` was
  pinned to a specific version, while others had wider version
specifiers. This lead to situations where incompatible versions were
installed.

To address these issues, this change is:

1.) Preventing silent build errors in `graylog-web-plugin` by to
  improving the `preinstall` task
2.) Pinning `babel` dependencies to constant, consistent versions