Allow user access to event annotaion for streams he has access to (#9762)
## Motivation
Prior to this change, the user needed additional access to `All Events`
and `All system events` to search with Event Annotation.

## Description
This change will change the streams used to check for permission.
Instead of effective streams which contains the Event Streams,
we use the streams the user has access to. We will filter with these
streams later the result of the Event Streams request.