Fixing initial height of widget container in edit mode by removing lazy loaded css.
Before this change the initial widget container `clientHeight` equaled the `scrollHeight` because the lazy loaded css defined `height: 100%` for the modal container.
We can remove this css file because we defined `height: 100%` for the modal content by default, the styling is not required and we will adjust / restyle the related components soon.