Style DataTableEntries for numeric values (#10195)
* Style DataTableEntries for numeric values

## Motivation
Prior to this change, numeric values where render just like the rest of
field types. But this makes it increasingly hard to read the values,
especially if the numbers are getting bigger. In a first step
@dennisoelkers introduced digig grouping.

## Description
This change will use monospace font for numeric values and align the
text on the right side so bigger numbers are easier to read.

* <th /> is no controll