Unset old group by form values when changing the type of the field
## Motivation
Prior to this change, the frontend included the old form values in the
request to the backend when applying a new config. For example if the
field type changed from time to value, we do not need the interval
values anymore but only the limit form value.

## Description
This change will set the old values to undefined before setting the new
default values.