Fixing loading lifecycle for exported dashboards with failing load hooks.
This change is making sure that a dashboard which has been exported from
a search loads properly if one of the (load/execute) hooks fails. There
were two issues which are fixed in this change:

  1. The view was immediately recreated in the `NewDashboardPage` after
     taken from state, due to a rerendering of the `NewDashboardPage`
  2. The `HookComponent` returned by `useLoadView` was not cleared after
     a successful retry.

This change is fixing these issues, ensuring that a search exported to a
dashboard loads fine, e.g. if containing parameters.