Cleanup message export format filter. (#10344)
This change is performing some cleanup for the message export format
filter. It is not a named filter anymore, because it requires to be run
at a pre matching stage in order to mutate the incoming request.
Therefore the `@SupportOnlyValidExportFormats` is removed for both the
filter and the `MessagesResource` REST class.