Hide save and load controls when editing a widget
## Motivation
Prior to this change, a page which was displaying a widget in edit mode,
the page was rendering two Save buttons. Also the complexity of states
to handle on each of the controls was not worth the gain.

## Description
This change will hide the view controls on dashboard and the saved
search controls on a search.