Do not display group by section in aggregation builder if widget has no groupings..
Before this change the group by section was visible in the widget edit mode, for example after creating a new widget. The section was displayed because the form value for `groupBy` was not empty.
With this change we are ensuring that `fromConfig` of the group by element returns `undefined` if no grouping exists.