Execute search only for focused widget
## Motivation
Prior to this change, whenever a search was executed the search was
executed for all widgets, regardless if they are seen or not.

## Description
This change will add a SearchAction which sets a filter by which the
search types will be filtered before beeing send to the server.
This filter will be set from the widget focus provider when ever a
widget gets focused or edited.