Fix case inconsistencies with ES cluster health status (#10428)
* Use consistent case for ES cluster health status

The cluster health status on ES6 was a lowercase string but on ES7 is
uppercase. This change transforms the status in ES7 into a lowercase
string.

* Make IndexerClusterHealthSummary case independent

The component made comparisons of ES cluster health status in a case
dependent way, causing #10407. This change ensures all comparisons
are done in a case independent way.

Fixes #10407