Track initial and mapped store state separately in `useStore`.
This change is modifying `useStore` to track the store state and its
mapped version separately. This allows us to declare an effect that is
solely registering a listen handler to the supplied store, regardless of
the props mapper function. Also, it allows us to reflect changes of the
supplied props mapper function (e.g. if it's passed as an arrow
function, so it's changing for each `useStore` call) without having
to reregister to the store. This is also avoiding a race condition where
reregistration to the store shadows an intermittent state update of the
store which is then not propagated to this `useStore` instance.